+++ example/CO_driver_blank.c
@@ -27,7 +27,6 @@
 
 
 #include "301/CO_driver.h"
-#include "301/CO_Emergency.h"
 
 
 /******************************************************************************/
@@ -67,13 +66,13 @@ CO_ReturnError_t CO_CANmodule_init(
     CANmodule->rxSize = rxSize;
     CANmodule->txArray = txArray;
     CANmodule->txSize = txSize;
+    CANmodule->CANerrorStatus = 0;
     CANmodule->CANnormal = false;
     CANmodule->useCANrxFilters = (rxSize <= 32U) ? true : false;/* microcontroller dependent */
     CANmodule->bufferInhibitFlag = false;
     CANmodule->firstCANtxMessage = true;
     CANmodule->CANtxCount = 0U;
     CANmodule->errOld = 0U;
-    CANmodule->em = NULL;
 
     for(i=0U; i<rxSize; i++){
         rxArray[i].ident = 0U;
@@ -196,7 +195,7 @@ CO_ReturnError_t CO_CANsend(CO_CANmodule_t *CANmodule, CO_CANtx_t *buffer){
     if(buffer->bufferFull){
         if(!CANmodule->firstCANtxMessage){
             /* don't set error, if bootup message is still on buffers */
-            CO_errorReport((CO_EM_t*)CANmodule->em, CO_EM_CAN_TX_OVERFLOW, CO_EMC_CAN_OVERRUN, buffer->ident);
+            CANmodule->CANerrorStatus |= CO_CAN_ERRTX_OVERFLOW;
         }
         err = CO_ERROR_TX_OVERFLOW;
     }
@@ -249,66 +248,62 @@ void CO_CANclearPendingSyncPDOs(CO_CANmodule_t *CANmodule){
 
 
     if(tpdoDeleted != 0U){
-        CO_errorReport((CO_EM_t*)CANmodule->em, CO_EM_TPDO_OUTSIDE_WINDOW, CO_EMC_COMMUNICATION, tpdoDeleted);
+        CANmodule->CANerrorStatus |= CO_CAN_ERRTX_PDO_LATE;
     }
 }
 
 
 /******************************************************************************/
-void CO_CANverifyErrors(CO_CANmodule_t *CANmodule){
-    uint16_t rxErrors, txErrors, overflow;
-    CO_EM_t* em = (CO_EM_t*)CANmodule->em;
-    uint32_t err;
+/* Get error counters from the module. If necessary, function may use
+    * different way to determine errors. */
+static uint16_t rxErrors=0, txErrors=0, overflow=0;
 
-    /* get error counters from module. Id possible, function may use different way to
-     * determine errors. */
-    rxErrors = CANmodule->txSize;
-    txErrors = CANmodule->txSize;
-    overflow = CANmodule->txSize;
+void CO_CANmodule_process(CO_CANmodule_t *CANmodule) {
+    uint32_t err;
 
     err = ((uint32_t)txErrors << 16) | ((uint32_t)rxErrors << 8) | overflow;
 
-    if(CANmodule->errOld != err){
+    if (CANmodule->errOld != err) {
+        uint16_t status = CANmodule->CANerrorStatus;
+
         CANmodule->errOld = err;
 
-        if(txErrors >= 256U){                               /* bus off */
-            CO_errorReport(em, CO_EM_CAN_TX_BUS_OFF, CO_EMC_BUS_OFF_RECOVERED, err);
+        if (txErrors >= 256U) {
+            /* bus off */
+            status |= CO_CAN_ERRTX_BUS_OFF;
         }
-        else{                                               /* not bus off */
-            CO_errorReset(em, CO_EM_CAN_TX_BUS_OFF, err);
+        else {
+            /* recalculate CANerrorStatus, first clear some flags */
+            status &= 0xFFFF ^ (CO_CAN_ERRTX_BUS_OFF |
+                                CO_CAN_ERRRX_WARNING | CO_CAN_ERRRX_PASSIVE |
+                                CO_CAN_ERRTX_WARNING | CO_CAN_ERRTX_PASSIVE);
 
-            if((rxErrors >= 96U) || (txErrors >= 96U)){     /* bus warning */
-                CO_errorReport(em, CO_EM_CAN_BUS_WARNING, CO_EMC_NO_ERROR, err);
-            }
-
-            if(rxErrors >= 128U){                           /* RX bus passive */
-                CO_errorReport(em, CO_EM_CAN_RX_BUS_PASSIVE, CO_EMC_CAN_PASSIVE, err);
-            }
-            else{
-                CO_errorReset(em, CO_EM_CAN_RX_BUS_PASSIVE, err);
-            }
-
-            if(txErrors >= 128U){                           /* TX bus passive */
-                if(!CANmodule->firstCANtxMessage){
-                    CO_errorReport(em, CO_EM_CAN_TX_BUS_PASSIVE, CO_EMC_CAN_PASSIVE, err);
-                }
-            }
-            else{
-                bool_t isError = CO_isError(em, CO_EM_CAN_TX_BUS_PASSIVE);
-                if(isError){
-                    CO_errorReset(em, CO_EM_CAN_TX_BUS_PASSIVE, err);
-                    CO_errorReset(em, CO_EM_CAN_TX_OVERFLOW, err);
-                }
+            /* rx bus warning or passive */
+            if (rxErrors >= 128) {
+                status |= CO_CAN_ERRRX_WARNING | CO_CAN_ERRRX_PASSIVE;
+            } else if (rxErrors >= 96) {
+                status |= CO_CAN_ERRRX_WARNING;
             }
 
-            if((rxErrors < 96U) && (txErrors < 96U)){       /* no error */
-                CO_errorReset(em, CO_EM_CAN_BUS_WARNING, err);
+            /* tx bus warning or passive */
+            if (txErrors >= 128) {
+                status |= CO_CAN_ERRTX_WARNING | CO_CAN_ERRTX_PASSIVE;
+            } else if (rxErrors >= 96) {
+                status |= CO_CAN_ERRTX_WARNING;
+            }
+
+            /* if not tx passive clear also overflow */
+            if ((status & CO_CAN_ERRTX_PASSIVE) == 0) {
+                status &= 0xFFFF ^ CO_CAN_ERRTX_OVERFLOW;
             }
         }
 
-        if(overflow != 0U){                                 /* CAN RX bus overflow */
-            CO_errorReport(em, CO_EM_CAN_RXB_OVERFLOW, CO_EMC_CAN_OVERRUN, err);
+        if (overflow != 0) {
+            /* CAN RX bus overflow */
+            status |= CO_CAN_ERRRX_OVERFLOW;
         }
+
+        CANmodule->CANerrorStatus = status;
     }
 }